cinema-ticketing/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
alembic
psycopg2-binary
mysql-connector-python
